requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
compressed_rtf==1.0.6
cryptography==42.0.8
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.28.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
IMAPClient==2.1.0
jiter==0.5.0
lxml==5.3.1
numpy==1.26.4
olefile==0.47
openai==1.40.0
opencv-python>=4.8
opencv-python-headless>=4.8
openparse==0.5.7
openpyxl==3.1.5
packaging==24.2
pdf2image==1.17.0
pdfminer.six>=20200401
pdfplumber==0.11.2
pillow==10.4.0
pip==24.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-pptx==0.6.23
regex==2024.7.24
requests==2.32.3
setuptools==69.5.1
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
tabulate==0.9.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.2.2
wheel==0.43.0
xlrd==1.2.0
XlsxWriter==3.2.2
python-docx
easyocr